+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
+## 1.0.11 - 2022-01-04
+
 ### Added
 - Allow preview current edits on [indoor=](https://indoorequal.org/) (thanks to François de Metz)